refactor: move files to `utils` and `hooks` to clean up pages folder
