+++ src/utils/generatePixCode.ts
